+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+
+* Fixed `--pass`, `--restart-if`, and `--restart-if-not` not allowing a leading negative number.
+
 ## v1.1.0 (2022-01-18)
 
 * Added version to executable properties.